# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ExportCreate": {
        "properties": {
          "activity_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Ids"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          }
        },
        "required": [
          "course_id"
        ],
        "title": "ExportCreate",
        "type": "object"
      },
      "ExportErrorRead": {
        "properties": {
          "file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "id",
          "message"
        ],
        "title": "ExportErrorRead",
        "type": "object"
      },
      "ExportFileRead": {
        "properties": {
          "activity_id": {
            "title": "Activity Id",
            "type": "string"
          },
          "activity_name": {
            "title": "Activity Name",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "export_mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Export Mime Type"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "output_path": {
            "title": "Output Path",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "student_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Email"
          },
          "student_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Name"
          }
        },
        "required": [
          "id",
          "activity_id",
          "activity_name",
          "student_email",
          "student_name",
          "source_name",
          "mime_type",
          "export_mime_type",
          "output_path",
          "status"
        ],
        "title": "ExportFileRead",
        "type": "object"
      },
      "ExportJobRead": {
        "properties": {
          "completed_files": {
            "title": "Completed Files",
            "type": "integer"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "course_name": {
            "title": "Course Name",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExportErrorRead"
            },
            "title": "Errors",
            "type": "array"
          },
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExportFileRead"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ExportStatus"
          },
          "total_files": {
            "title": "Total Files",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "course_id",
          "course_name",
          "status",
          "total_files",
          "completed_files"
        ],
        "title": "ExportJobRead",
        "type": "object"
      },
      "ExportStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "ExportStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Classroom Downloader API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/exports": {
      "post": {
        "operationId": "create_export_api_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportJobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Export"
      }
    },
    "/api/exports/{job_id}": {
      "get": {
        "operationId": "read_export_api_exports__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportJobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Export"
      }
    },
    "/api/exports/{job_id}/files/{file_id}/content": {
      "get": {
        "operationId": "stream_export_file_api_exports__job_id__files__file_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "title": "File Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Export File"
      }
    },
...
  }
}